feat: upgrade to latest mongoose

+++ test/utils/dropModel.ts
@@ -1,5 +1,7 @@
 import mongoose from 'mongoose';
 
 export default async function dropModel(modelName: string): Promise<void> {
-	await mongoose.connection.db.dropCollection(`mongoose_delete_${modelName}`);
+	if (mongoose.connection.db) {
+		await mongoose.connection.db.dropCollection(`mongoose_delete_${modelName}`);
+	}
 }
